Moved files to specific folders according the feature
