+++ app/src/main/AndroidManifest.xml
@@ -14,10 +14,10 @@
         android:theme="@style/Theme.ValorantPruebaApi"
         tools:targetApi="31">
         <activity
-            android:name=".ActivityMapUnit"
+            android:name=".maps.ActivityMapUnit"
             android:exported="false" />
         <activity
-            android:name=".ActivityMaps"
+            android:name=".maps.ActivityMaps"
             android:exported="false" />
         <activity
             android:name=".RegisterScreen"
@@ -26,10 +26,10 @@
             android:name=".MainScreen"
             android:exported="false" />
         <activity
-            android:name=".ActivityAgentUnit"
+            android:name=".agents.ActivityAgentUnit"
             android:exported="false" />
         <activity
-            android:name=".ActivityAgents"
+            android:name=".agents.ActivityAgents"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
