+++ app/src/main/java/com/example/valorantpruebaapi/agents/ActivityAgents.kt
@@ -1,4 +1,4 @@
-package com.example.valorantpruebaapi
+package com.example.valorantpruebaapi.agents
 
 import ClasesApi.*
 import android.content.Intent
@@ -6,7 +6,6 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import androidx.recyclerview.widget.GridLayoutManager
-import com.example.valorantpruebaapi.adapter.AgentAdapter
 import com.example.valorantpruebaapi.databinding.ActivityAgentsBinding
 import retrofit2.Call
 import retrofit2.Callback
