+++ app/src/main/res/layout/activity_agent_unit.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ActivityAgentUnit">
+    tools:context=".agents.ActivityAgentUnit">
 
     <ImageView
         android:id="@+id/imageViewFondo"
