+++ app/src/main/res/layout/activity_agents.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ActivityAgents">
+    tools:context=".agents.ActivityAgents">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewAgents"
