+++ app/src/main/res/layout/activity_maps.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ActivityMaps">
+    tools:context=".maps.ActivityMaps">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewMaps"
